Several Magic::Events using reinterpret_cast now use static_cast - compiles on Windows

+++ gui/include/gui/windows/client.hpp
@@ -58,7 +58,7 @@ namespace GUI {
         };
 
         void client1(
-            const int i,
+            const size_t i,
             ImGuiID center_id,
             Client &client,
             MenuBarEnables &enables,
